add middleware/controller
