+++ app.js
@@ -2,7 +2,7 @@ const express = require('express')
 const logger = require('morgan')
 const cors = require('cors')
 
-const contactsRouter = require('./routes/api/contacts')
+const contactsRouter = require('./src/routes/api/contacts')
 
 const app = express()
